attendance_sync/requirements.txt
mysql-connector-python
supabase
python-dotenv